# doc/swagger_v1.yaml
openapi: 3.0.0
info:
  title: GeoControl API
  description: |
    **GeoControl System Overview**

    GeoControl is a software system designed for monitoring physical and environmental variables in various contexts: from hydrogeological analyses of mountain areas to the surveillance of historical buildings, and even the control of internal parameters (such as temperature or lighting) in residential or working environments.  
    It was commissioned by the Union of Mountain Communities of the Piedmont region for managing the hydrogeological state of its territories. However, thanks to its modular structure, GeoControl has also been commercialized to different public or private entities requiring continuous monitoring of physical parameters.  
    The system meets high reliability requirements and must ensure that no more than six measurements per year, per sensor, are lost. From an organizational perspective, GeoControl uses a token-based authentication mechanism, with different user roles to regulate access levels to the functionalities.

    ---

    **Hierarchical Structure and Components**

    GeoControl’s hierarchy is organized into three main levels:

    1. **Network**  
       - A logical grouping that acts as a container for multiple gateways (and their associated sensors).  
       - Identified by a unique alphanumeric code (chosen at creation time) and can represent, for example, a monitoring network for a single municipality or an entire building.  
...

...

...

...

    ---

...

...
servers:
...
paths:
...
  /networks:
    get:
      summary: Retrieve all networks
      description: This endpoint can be used by any kind of authenticated user
      tags:
        - Networks
      security:
        - bearerAuth: []
      responses:
        "200":
          description: A list of networks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Network"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDTO"
              example:
                code: "401"
                name: "UnauthorizedError"
                message: "Unauthorized: Invalid token format"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDTO"
              example:
                code: "500"
                name: "InternalServerError"
                message: "Internal server error"
    post:
      summary: Create a new network
      description: |
        This endpoint can be used only by Admin and Operator users.
        Each request to create a Network is considered valid even if it contains nested Gateways and Sensors. However, only the Network data will be processed. Any nested objects included in the request will be ignored and not created through this operation.
      tags:
        - Networks
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code:
                  type: string
                  minLength: 1
                  description: Alphanumeric unique identifier for the network
                  example: NET01
                name:
                  type: string
                  description: Name of the network
                  example: Alp Monitor
                description:
                  type: string
                  description: Description of the network
                  example: Alpine Weather Monitoring Network
                gateways:
                  type: array
                  items:
                    $ref: "#/components/schemas/Gateway"
            example:
              code: "NET01"
              name: "Alp Monitor"
              description: "Alpine Weather Monitoring Network"
      responses:
        "201":
          description: Network created
        "400":
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDTO"
              example:
                code: "400"
                name: "BadRequest"
                message: "/body/xxx' must have required property 'xxx'"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDTO"
              example:
                code: "401"
                name: "UnauthorizedError"
                message: "Unauthorized: Invalid token format"
        "403":
          description: Insufficient rights
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDTO"
              example:
                code: "403"
                name: "InsufficientRightsError"
                message: "Forbidden: Insufficient rights"
        "409":
          description: Network code already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDTO"
              example:
                code: "409"
                name: "ConflictError"
                message: "Entity with code xxxxx already exists"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDTO"
              example:
                code: "500"
                name: "InternalServerError"
                message: "Internal server error"
  /networks/{networkCode}:
    get:
      summary: Retrieve a specific network
      description: This endpoint can be used by any kind of authenticated user
      tags:
        - Networks
      security:
        - bearerAuth: []
      parameters:
        - name: networkCode
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: A single network
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Network"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDTO"
              example:
                code: "401"
                name: "UnauthorizedError"
                message: "Unauthorized: Invalid token format"
        "404":
          description: Network not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDTO"
              example:
                code: "404"
                name: "NotFoundError"
                message: "Entity not found"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDTO"
              example:
                code: "500"
                name: "InternalServerError"
                message: "Internal server error"
    patch:
      summary: Update a network
      description: |
        This endpoint can be used only by Admin and Operator users.
        Each request to update a Network is considered valid even if it contains nested Gateways and Sensors. However, only the Network data will be processed. Any nested objects included in the request will be ignored and not created or updated through this operation. The request must allow the update of all the Network fields, including the Network code.
      tags:
        - Networks
      security:
        - bearerAuth: []
      parameters:
        - name: networkCode
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Network"
            example:
              code: "NEWCODE"
              name: "Updated Name"
              description: "Updated Description"
      responses:
        "204":
          description: Network updated
        "400":
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDTO"
              example:
                code: "400"
                name: "BadRequest"
                message: "/body/xxx' must have required property 'xxx'"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDTO"
              example:
                code: "401"
                name: "UnauthorizedError"
                message: "Unauthorized: Invalid token format"
        "403":
          description: Insufficient rights
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDTO"
              example:
                code: "403"
                name: "InsufficientRightsError"
                message: "Forbidden: Insufficient rights"
        "404":
          description: Network not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDTO"
              example:
                code: "404"
                name: "NotFoundError"
                message: "Entity not found"
        "409":
          description: Network code already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDTO"
              example:
                code: "409"
                name: "ConflictError"
                message: "Entity with code xxxxx already exists"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDTO"
              example:
                code: "500"
                name: "InternalServerError"
                message: "Internal server error"
    delete:
      summary: Delete a network
      description: This endpoint can be used only by Admin and Operator users
      tags:
        - Networks
      security:
        - bearerAuth: []
      parameters:
        - name: networkCode
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Network deleted
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDTO"
              example:
                code: "401"
                name: "UnauthorizedError"
                message: "Unauthorized: Invalid token format"
        "403":
          description: Insufficient rights
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDTO"
              example:
                code: "403"
                name: "InsufficientRightsError"
                message: "Forbidden: Insufficient rights"
        "404":
          description: Network not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDTO"
              example:
                code: "404"
                name: "NotFoundError"
                message: "Entity not found"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDTO"
              example:
                code: "500"
                name: "InternalServerError"
                message: "Internal server error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Network:
      type: object
      properties:
        code:
          type: string
          minLength: 1
          description: Alphanumeric unique identifier for the network
          example: NET01
        name:
          type: string
          description: Name of the network
          example: Alp Monitor
        description:
          type: string
          description: Description of the network
          example: Alpine Weather Monitoring Network
        gateways:
          type: array
          items:
            $ref: "#/components/schemas/Gateway"
    Gateway:
      type: object
      properties:
        macAddress:
          type: string
          minLength: 1
          description: Unique identifier for the gateway (MAC address)
          example: 94:3F:BE:4C:4A:79
        name:
          type: string
          description: Name of the gateway
          example: GW01
        description:
          type: string
          description: Description of the gateway
          example: on-field aggregation node
        sensors:
          type: array
          items:
            $ref: "#/components/schemas/Sensor"
    Sensor:
      type: object
      properties:
        macAddress:
          type: string
          minLength: 1
          description: Unique identifier for the sensor (MAC address)
          example: 71:B1:CE:01:C6:A9
        name:
          type: string
          description: Name of the sensor
          example: TH01
        description:
          type: string
          description: Description of the sensor
          example: External thermometer
        variable:
          type: string
          description: Variable associated with the sensor
          example: temperature
        unit:
          type: string
          description: Unit of the variable
          example: C
...
    ErrorDTO:
      type: object
      required:
        - code
      properties:
        code:
          type: number
          description: HTTP code of the error
          example: 404
        name:
          type: string
          description: Name of the error
          example: NotFoundError
        message:
          type: string
          description: Message describing the type and the cause of the error
          example: Entity not found